fluent 1.4 w/ connection pool
